final project feedback
[This comment is a supplement to the feedback in email.]

Generally a pull request is meant to be accepted into a code base. This one should not be. I’m including code changes that are improvements, but also fragments and comments that are meant to be read as commentary, but not added to the repo.

You were only scored on the material on the course policies page. I’ve included many additional comments and suggestions for further improving your code quality beyond what this course asks for, or sometimes just as proposals for areas you may want to look at. (Generally, comments that include “or” or “you could also” are alternatives as a matter of style preference, not necessarily improvements.)

General comments on the code and organization:
- Considerable code duplication, use of magic numbers.
- Yay on having test files!
- A convention for data files is to collect them into a data or test or test/data subdirectory.
- Dead code (commented-out, unfinished and uncalled) code should generally be removed
  - Once you’re more confident in your git fu, you can place old code in a branch; or just delete it but git tag a commit that still contains any code you might want to refer back to
  - I believe sa_cache_BACKUP_65172.txt, etc., are not used, and are an artifact of a git merge? They should not be in the project.
  - You’ve got several *.png files – place them in a subdirectory
- As a binary file, choose another file extension that “.txt” for sa_cache.txt.